requirements.txt
Flask
SQLAlchemy
Pillow
oauth2client
requests
passlib